move: 파일 경로 변경
info -> member
