--- src/main/java/com/atwoz/member/exception/exceptions/info/hobby/HobbySizeException.java
@@ -1,8 +1,0 @@
-package com.atwoz.member.exception.exceptions.info.hobby;
-
-public class HobbySizeException extends RuntimeException {
-
-    public HobbySizeException() {
-        super("선택할 수 있는 취미 갯수와 맞지 않습니다.");
-    }
-}
