--- src/main/java/com/atwoz/member/exception/exceptions/info/option/InvalidMbtiException.java
@@ -1,8 +1,0 @@
-package com.atwoz.member.exception.exceptions.info.option;
-
-public class InvalidMbtiException extends RuntimeException {
-
-    public InvalidMbtiException() {
-        super("등록되지 않은 MBTI입니다.");
-    }
-}
