--- src/main/java/com/atwoz/member/exception/exceptions/info/option/InvalidReligionException.java
@@ -1,8 +1,0 @@
-package com.atwoz.member.exception.exceptions.info.option;
-
-public class InvalidReligionException extends RuntimeException {
-
-    public InvalidReligionException() {
-        super("등록되지 않은 종교입니다.");
-    }
-}
