--- src/main/java/com/atwoz/member/exception/exceptions/info/option/InvalidSmokeException.java
@@ -1,8 +1,0 @@
-package com.atwoz.member.exception.exceptions.info.option;
-
-public class InvalidSmokeException extends RuntimeException {
-
-    public InvalidSmokeException() {
-        super("등록되지 않은 흡연 단계입니다.");
-    }
-}
