--- src/main/java/com/atwoz/member/exception/exceptions/info/profile/body/AgeRangeException.java
@@ -1,8 +1,0 @@
-package com.atwoz.member.exception.exceptions.info.profile.body;
-
-public class AgeRangeException extends RuntimeException {
-
-    public AgeRangeException() {
-        super("나이 범위가 올바르지 않습니다.");
-    }
-}
