--- src/main/java/com/atwoz/member/exception/exceptions/info/profile/body/HeightRangeException.java
@@ -1,8 +1,0 @@
-package com.atwoz.member.exception.exceptions.info.profile.body;
-
-public class HeightRangeException extends RuntimeException {
-
-    public HeightRangeException() {
-        super("키 범위가 올바르지 않습니다.");
-    }
-}
