--- src/main/java/com/atwoz/member/exception/exceptions/info/profile/job/InvalidJobException.java
@@ -1,8 +1,0 @@
-package com.atwoz.member.exception.exceptions.info.profile.job;
-
-public class InvalidJobException extends RuntimeException {
-
-    public InvalidJobException() {
-        super("직업을 찾을 수 없습니다.");
-    }
-}
