--- src/main/java/com/atwoz/member/exception/exceptions/info/style/InvalidStyleException.java
@@ -1,8 +1,0 @@
-package com.atwoz.member.exception.exceptions.info.style;
-
-public class InvalidStyleException extends RuntimeException {
-
-    public InvalidStyleException() {
-        super("등록되지 않은 스타일입니다.");
-    }
-}
