--- src/main/java/com/atwoz/member/exception/exceptions/info/style/StyleDuplicateException.java
@@ -1,8 +1,0 @@
-package com.atwoz.member.exception.exceptions.info.style;
-
-public class StyleDuplicateException extends RuntimeException {
-
-    public StyleDuplicateException() {
-        super("스타일을 중복 선택하면 안 됩니다.");
-    }
-}
