+++ src/main/java/com/atwoz/member/exception/exceptions/member/profile/HobbyDuplicateException.java
@@ -1,0 +1,8 @@
+package com.atwoz.member.exception.exceptions.member.profile;
+
+public class HobbyDuplicateException extends RuntimeException {
+
+    public HobbyDuplicateException() {
+        super("취미를 중복 선택하면 안 됩니다.");
+    }
+}
