+++ src/main/java/com/atwoz/member/exception/exceptions/member/profile/InvalidDrinkException.java
@@ -1,0 +1,8 @@
+package com.atwoz.member.exception.exceptions.member.profile;
+
+public class InvalidDrinkException extends RuntimeException {
+
+    public InvalidDrinkException() {
+        super("등록되지 않은 음주 단계입니다.");
+    }
+}
