+++ src/main/java/com/atwoz/member/exception/exceptions/member/profile/InvalidHobbyException.java
@@ -1,0 +1,8 @@
+package com.atwoz.member.exception.exceptions.member.profile;
+
+public class InvalidHobbyException extends RuntimeException {
+
+    public InvalidHobbyException() {
+        super("등록되지 않은 취미입니다.");
+    }
+}
