+++ src/main/java/com/atwoz/member/exception/exceptions/member/profile/InvalidJobException.java
@@ -1,0 +1,8 @@
+package com.atwoz.member.exception.exceptions.member.profile;
+
+public class InvalidJobException extends RuntimeException {
+
+    public InvalidJobException() {
+        super("직업을 찾을 수 없습니다.");
+    }
+}
