+++ src/main/java/com/atwoz/member/exception/exceptions/member/profile/InvalidReligionException.java
@@ -1,0 +1,8 @@
+package com.atwoz.member.exception.exceptions.member.profile;
+
+public class InvalidReligionException extends RuntimeException {
+
+    public InvalidReligionException() {
+        super("등록되지 않은 종교입니다.");
+    }
+}
