+++ src/main/java/com/atwoz/member/exception/exceptions/member/profile/InvalidSmokeException.java
@@ -1,0 +1,8 @@
+package com.atwoz.member.exception.exceptions.member.profile;
+
+public class InvalidSmokeException extends RuntimeException {
+
+    public InvalidSmokeException() {
+        super("등록되지 않은 흡연 단계입니다.");
+    }
+}
