+++ src/main/java/com/atwoz/member/exception/exceptions/member/profile/StyleDuplicateException.java
@@ -1,0 +1,8 @@
+package com.atwoz.member.exception.exceptions.member.profile;
+
+public class StyleDuplicateException extends RuntimeException {
+
+    public StyleDuplicateException() {
+        super("스타일을 중복 선택하면 안 됩니다.");
+    }
+}
