+++ src/main/java/com/atwoz/member/exception/exceptions/member/profile/contact/LatitudeRangeException.java
@@ -1,0 +1,8 @@
+package com.atwoz.member.exception.exceptions.member.profile.contact;
+
+public class LatitudeRangeException extends RuntimeException {
+
+    public LatitudeRangeException() {
+        super("위도 범위가 올바르지 않습니다.");
+    }
+}
