+++ src/main/java/com/atwoz/member/exception/exceptions/member/profile/contact/LongitudeRangeException.java
@@ -1,0 +1,8 @@
+package com.atwoz.member.exception.exceptions.member.profile.contact;
+
+public class LongitudeRangeException extends RuntimeException {
+
+    public LongitudeRangeException() {
+        super("경도 범위가 올바르지 않습니다.");
+    }
+}
