+++ src/main/java/com/atwoz/member/exception/exceptions/member/profile/physical_profile/HeightRangeException.java
@@ -1,0 +1,8 @@
+package com.atwoz.member.exception.exceptions.member.profile.physical_profile;
+
+public class HeightRangeException extends RuntimeException {
+
+    public HeightRangeException() {
+        super("키 범위가 올바르지 않습니다.");
+    }
+}
