+++ src/main/java/com/atwoz/member/infrastructure/member/YearManagerImpl.java
@@ -1,0 +1,15 @@
+package com.atwoz.member.infrastructure.member;
+
+import com.atwoz.member.domain.member.profile.physical_profile.YearManager;
+import java.time.LocalDateTime;
+import org.springframework.stereotype.Component;
+
+@Component
+public class YearManagerImpl implements YearManager {
+
+    @Override
+    public int getCurrentYear() {
+        LocalDateTime now = LocalDateTime.now();
+        return now.getYear();
+    }
+}
