+++ src/test/java/com/atwoz/member/infrastructure/member/physical_profile/FakeYearManager.java
@@ -1,0 +1,13 @@
+package com.atwoz.member.infrastructure.member.physical_profile;
+
+import com.atwoz.member.domain.member.profile.physical_profile.YearManager;
+
+public class FakeYearManager implements YearManager {
+
+    private static final int YEAR = 2024;
+
+    @Override
+    public int getCurrentYear() {
+        return YEAR;
+    }
+}
